// src/app/models/checklist.ts
export class Checklist{
    JobCardCheckListID: number;
    JobCardID: number;
    CheckListID: number;
    StatusID: number;
    CheckListTypeID: number;
    IsModified: boolean;
    IsDeleted: number;
    Note: string;
    CreatedBy: number;
    ModifiedBy: number;
    IsChecked: boolean;
    AppointmentID: number;
} 